Remove separate test file for mmol/L - simplifies merging for others
